// src/API.ts
export type Todo = {
  __typename: "Todo",
  id: string,
  name: string,
  description?: string | null,
  image?: string | null,
  owner: string,
  createdAt: string,
  updatedAt: string,
};
// ...
export type ModelTodoConnection = {
  __typename: "ModelTodoConnection",
  items:  Array<Todo | null >,
  nextToken?: string | null,
};
// ...
export type ListTodosQuery = {
  listTodos?:  {
    __typename: "ModelTodoConnection",
    items:  Array< {
      __typename: "Todo",
      id: string,
      name: string,
      description?: string | null,
      image?: string | null,
      owner: string,
      createdAt: string,
      updatedAt: string,
    } | null >,
    nextToken?: string | null,
  } | null,
};